backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
lxml==4.9.3
python-docx==0.8.11
scikit-learn==1.3.2
numpy>=1.21.0,<1.25.0
pandas>=2.0.0,<2.2.0
nltk==3.8.1
werkzeug==3.0.1
matplotlib>=3.5.0
